db: move pack/unpack serialization code into new db_util.py
